requirements.txt
streamlit
pdfplumber
pdf417gen
Pillow
pytesseract
google-genai
